Move pipes to their own directory
