+++ src/app/players/pipes/prettier-player-name.pipe.ts
